+++ pages/edit-quest/_id/index.vue
@@ -237,6 +237,7 @@ import { ERC20, WorkQuest, WQPromotion } from '~/abi';
 import { error, success } from '~/utils/web3';
 import { CommissionForCreatingAQuest } from '~/utils/сonstants/commission';
 import { images } from '~/utils/images';
+import walletOperations from '~/plugins/mixins/walletOperations';
 
 const { GeoCode } = require('geo-coder');
 
@@ -244,6 +245,7 @@ export default {
   name: 'EditQuest',
   EditQuestState,
   middleware: 'employer-role',
+  mixins: [walletOperations],
   data() {
     return {
       period: 0,
